user done with password handling

+++ application/Http/Controllers/UserController.php
@@ -5,9 +5,12 @@ namespace Intranet\Http\Controllers;
 
 
 use Core\Contracts\Http\RequestInterface;
+use Core\Contracts\Http\ResponseInterface;
 use Intranet\Contracts\Auth\AuthInterface;
+use Intranet\Contracts\Repositories\RoleRepositoryInterface;
 use Intranet\Contracts\Repositories\UserRepositoryInterface;
 use Intranet\Services\Auth\Auth;
+use Intranet\Services\Mail\Mail;
 
 class UserController
 {
@@ -17,39 +20,50 @@ class UserController
     private $userRepository;
 
     /**
+     * @var RoleRepositoryInterface
+     */
+    private $roleRepository;
+
+    /**
      * @var Auth
      */
     private $auth;
+    /**
+     * @var Mail
+     */
+    private $mail;
 
     /**
      * UserController constructor.
      * @param UserRepositoryInterface $userRepository
+     * @param RoleRepositoryInterface $roleRepository
      * @param AuthInterface $auth
-     * @internal param ResponseFactoryInterface $responseFactory
-     * @internal param ResponseInterface $response
+     * @param Mail $mail
      */
-    public function __construct(UserRepositoryInterface $userRepository, AuthInterface $auth)
+    public function __construct(UserRepositoryInterface $userRepository, RoleRepositoryInterface $roleRepository, AuthInterface $auth, Mail $mail)
     {
         $this->userRepository = $userRepository;
         $this->auth = $auth;
+        $this->roleRepository = $roleRepository;
+        $this->mail = $mail;
     }
 
     /**
      * Create a new user.
      * @param RequestInterface $request
-     * @return \Core\Contracts\Http\ResponseInterface
+     * @return ResponseInterface
      */
     public function create(RequestInterface $request)
     {
         if (!$this->userRepository->hasPermission($this->auth->user()['username'], 'user_manage')) {
-            return \jsonError(403, \text('base.permission_denied'));
+            return \jsonError(403, ['permission' => [\text('base.permission_denied')]]);
         }
 
         $valid = $request->validate([
             'username' => [
                 'required',
                 'regex' => [
-                    'pattern' => "/[\w\.\-]{5,32}/"
+                    'pattern' => "/[A-Za-z0-9\_\-\.]{5,32}/"
                 ],
                 'unique' => [
                     'table' => \config('database.tables.user'),
@@ -69,12 +83,6 @@ class UserController
                 ],
                 'email'
             ],
-            'password' => [
-                'required',
-                'min_length' => [
-                    'min' => 6
-                ],
-            ],
             'role_name' => [
                 'required',
                 'exists' => [
@@ -93,30 +101,32 @@ class UserController
             'username' => $request->username,
             'name' => $request->name,
             'email' => $request->email,
-            'password' => $this->userRepository->hashPassword($request->password),
-            'role_name' => $request->role,
-            'password_reset_token' => null
+            'password' => null,
+            'role_name' => $request->role_name,
+            'password_reset_token' => null,
+            'password_reset_expire_at' => null
         ];
 
         $user = $this->userRepository->save($user);
-        unset($user['password']);
+        $user = $this->userRepository->toPublic($user);
 
-        return \json($user);
+        return \json(['user' => $user, 'message' => \text('app.user.create.success')]);
     }
 
     /**
      * Update existing user.
      * @param RequestInterface $request
-     * @return \Core\Contracts\Http\ResponseInterface
+     * @return ResponseInterface
      */
     public function update(RequestInterface $request)
     {
         $valid = $request->validate([
             'username' => [
                 'required',
-                'regex' => [
-                    'pattern' => "/[\w\.\-]{5,32}/"
-                ],
+                'exists' => [
+                    'table' => \config('database.tables.user'),
+                    'column' => 'username'
+                ]
             ],
             'name' => [
                 'max_length' => [
@@ -128,11 +138,6 @@ class UserController
                     'max' => 255,
                 ],
                 'email'
-            ],
-            'password' => [
-                'min_length' => [
-                    'min' => 6
-                ],
             ],
             'role_name' => [
                 'exists' => [
@@ -152,14 +157,17 @@ class UserController
         $canManage = $this->userRepository->hasPermission($loggedUsername, 'user_manage');
 
         if ($user['username'] != $loggedUsername and !$canManage) {
-            return \jsonError(403, \text('base.permission_denied'));
+            return \jsonError(403, ['permission' => [\text('base.permission_denied')]]);
         }
 
         $user['name'] = \is_null($request->name) ? $user['name'] : $request->name;
         $user['email'] = $request->email ?? $user['email'];
-        $user['password'] = $request->password ? $this->userRepository->hashPassword($request->password) : $user['password'];
 
-        if ($canManage) {
+        if ($canManage and $request->role_name) {
+            if ($user['username'] == $loggedUsername and $user['role_name'] == 'admin' and $request->role_name != 'admin') {
+                return \jsonError(403, ['role_name' => [\text('app.user.update.admin_role_denied')]]);
+            }
+
             $user['role_name'] = $request->role_name ?? $user['role_name'];
         }
 
@@ -172,12 +180,12 @@ class UserController
     /**
      * Delete the user.
      * @param RequestInterface $request
-     * @return \Core\Contracts\Http\ResponseInterface
+     * @return ResponseInterface
      */
     public function delete(RequestInterface $request)
     {
         if (!$this->userRepository->hasPermission($this->auth->user()['username'], 'user_manage')) {
-            return \jsonError(403, \text('base.permission_denied'));
+            return \jsonError(403, ['permission' => [\text('base.permission_denied')]]);
         }
 
         $valid = $request->validate([
@@ -195,30 +203,52 @@ class UserController
             return \jsonError(422, $errors);
         }
 
+        if ($request->username == $this->auth->user()['username']) {
+            return \jsonError(403, ['delete' => [\text('app.user.delete.delete_self_denied')]]);
+        }
+
         $this->userRepository->delete($request->username);
 
-        return \json();
+        return \json(['message' => \text('app.user.delete.success', ['username' => $request->username])]);
     }
 
     /**
      * Show page with listed users.
      * @param RequestInterface $request
-     * @return \Core\Contracts\Http\ResponseInterface
+     * @return ResponseInterface
      */
     public function list(RequestInterface $request)
     {
         if (!$this->userRepository->hasPermission($this->auth->user()['username'], "user_manage")) {
-            return \error(403, \text('base.permission_denied'));
+            return \error(403, ['permission' => [\text('base.permission_denied')]]);
         }
 
-        $users = $this->userRepository->findAll('name', false);
+        $users = $this->userRepository->findAll();
+        $roles = $this->roleRepository->findAll();
 
-        return \html('users', ['users' => $users]);
+        return \html('users', ['users' => $users, 'roles' => $roles]);
+    }
+
+    /**
+     * Show table with listed users.
+     * @param RequestInterface $request
+     * @return ResponseInterface
+     */
+    public function usersTable(RequestInterface $request)
+    {
+        if (!$this->userRepository->hasPermission($this->auth->user()['username'], "user_manage")) {
+            return \jsonError(403, ['permission' => [\text('base.permission_denied')]]);
+        }
+
+        $users = $this->userRepository->findAll();
+        $roles = $this->roleRepository->findAll();
+
+        return \html('components.users_table', ['users' => $users, 'roles' => $roles]);
     }
 
     /**
      * @param RequestInterface $request
-     * @return \Core\Contracts\Http\ResponseInterface|string
+     * @return ResponseInterface|string
      */
     public function showSettings(RequestInterface $request)
     {
@@ -260,7 +290,7 @@ class UserController
     /**
      * Change user password.
      * @param RequestInterface $request
-     * @return \Core\Contracts\Http\ResponseInterface|string
+     * @return ResponseInterface|string
      */
     public function changePassword(RequestInterface $request)
     {
@@ -272,25 +302,84 @@ class UserController
                     'column' => 'username'
                 ]
             ],
-            'token' => [
-                'required'
-            ]
+            'password' => [
+                'required',
+                'min_length' => [
+                    'min' => 6
+                ]
+            ],
         ]);
 
         if (!$valid) {
             $errors = $request->errors();
-            return \error(422, $errors);
+            return \jsonError(422, $errors);
         }
 
-        return \html('change_password', ['username' => $request->username, 'token' => $request->token]);
+        $user = $this->userRepository->findByUsername($request->username);
+
+        if ($request->_password) {
+            $hash = $user['password'];
+            if (!$this->userRepository->verifyPassword($request->_password, $hash)) {
+                return \jsonError(403, ['password' => [\text('app.user.change_password.invalid_password')]]);
+            }
+        }
+        else if ($request->token) {
+            $hash = $user['password_reset_token'];
+            $now = \strtotime('now');
+            $expire = \strtotime($user['password_expire_at']);
+
+            if (!$hash or !$expire or $now > $expire or !$this->userRepository->verifyPassword($request->token, $hash)){
+                return \jsonError(403, ['token' => [\text('app.user.change_password.invalid_token')]]);
+            }
+        }
+        else {
+            return \jsonError(403, ['auth' => [\text('app.user.change_password.no_auth')]]);
+        }
+
+        $user['password_reset_token'] = null;
+        $user['password'] = $this->userRepository->hashPassword($request->password);
+
+        $this->userRepository->save($user);
+
+        return \json(['message' => \text('app.user.change_password.success')]);
     }
 
     /**
      * Send change password email
      * @param RequestInterface $request
      */
-    public function resetPassword(RequestInterface $request)
+    public function sendChangePasswordEmail(RequestInterface $request)
     {
+        $valid = $request->validate([
+            'email' => [
+                'required',
+                'email'
+            ]
+        ]);
 
+        if (!$valid) {
+            $errors = $request->errors();
+            return \jsonError(422, $errors);
+        }
+
+        $user = $this->userRepository->findByEmail($request->email);
+
+        if ($user) {
+            if (!$user['password_reset_expire_at'] or \strtotime($user['password_reset_expire_at']) < \strtotime('now')) {
+                $token = \random_string(128);
+                $user['password_reset_token'] = $this->userRepository->hashPassword($token);
+                $user['password_reset_expire_at'] =  date("Y-m-d H:i:s", \strtotime('+3 hour'));
+
+                $this->userRepository->save($user);
+
+                $url = \sprintf("%s/user/change-password?username=%s&token=%s", \config('app.url'), $user['username'], $token);
+                $subject = \text('app.user.change_password.mail.subject');
+                $message = \text('app.user.change_password.mail.message', ['url' => $url]);
+
+                $this->mail->send($user['email'], $subject, $message);
+            }
+        }
+
+        return \json(['message' => \text('app.user.change_password.mail.sent')]);
     }
 }
